requirements.txt
requests==2.31.0
pandas
streamlit
pytest
python-dotenv
# flowise